ai/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
exceptiongroup==1.2.2
flatbuffers==25.1.24
fonttools==4.55.6
frozenlist==1.5.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.70.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imbalanced-learn==0.13.0
jax==0.5.0
jaxlib==0.5.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
keras==3.8.0
kiwisolver==1.4.8
langchain==0.3.19
langchain-core==0.3.37
langchain-text-splitters==0.3.6
langsmith==0.3.10
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mediapipe==0.10.20
ml-dtypes==0.4.1
multidict==6.1.0
namex==0.0.8
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.0
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pillow==11.1.0
propcache==0.3.0
protobuf==4.25.6
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
pypdf==5.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
scikit-learn==1.6.1
scipy==1.15.1
sentencepiece==0.2.0
six==1.17.0
sklearn-compat==0.1.3
sniffio==1.3.1
sounddevice==0.5.1
SQLAlchemy==2.0.38
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
threadpoolctl==3.5.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.18.3
zstandard==0.23.0
